+++ mcp-api-backend/api/v1/adminInfo.py
@@ -1,0 +1,13 @@
+from fastapi import APIRouter, Depends
+from sqlalchemy.orm import Session
+from db.connection import get_db
+from service import user_service
+
+router = APIRouter()
+
+@router.get("/")
+def get_admin_info(db: Session = Depends(get_db)):
+    admin = user_service.get_admin_info(db)
+
+    return admin
+
