+++ mcp-api-backend/api/v1/hello.py
@@ -1,0 +1,10 @@
+from fastapi import APIRouter
+
+router = APIRouter()
+
+@router.get("/")
+def hello_world():
+    """
+    API 설명
+    """
+    return "Hello World"
