+++ mcp-api-backend/api/v1/instance.py
@@ -1,0 +1,9 @@
+from fastapi import APIRouter
+from service import instance_service
+
+router = APIRouter()
+
+@router.get("/")
+def get_instance_state():
+    state = instance_service.get_instance_state()
+    return state
