+++ mcp-api-backend/db/session.py
@@ -1,0 +1,29 @@
+from config.database_config import settings
+import logging
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker, declarative_base
+
+logger = logging.getLogger("uvicorn")
+
+SQLALCHEMY_DATABASE_URL = "{}://{}:{}@{}:{}/{}".format(
+    settings.DB_ENGINE,
+    settings.DB_USERNAME,
+    settings.DB_PASS,
+    settings.DB_HOST,
+    settings.DB_PORT,
+    settings.DB_NAME,
+)
+
+engine = create_engine(
+    SQLALCHEMY_DATABASE_URL,
+    pool_pre_ping=True,
+    pool_recycle=3600,
+    pool_size=200,
+    max_overflow=-1,
+    echo=settings.DEBUG  # 쿼리 로그 출력
+)
+
+logger.info(f"데이터베이스 URL: {SQLALCHEMY_DATABASE_URL}")
+SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+
+Base = declarative_base()
