+++ mcp-api-backend/src/shared/security/deps.py
@@ -1,13 +1,3 @@
-from config.database_config import SessionLocal
-from pyparsing import Generator
-
 """
 의존성 관리 모듈
 """
-
-def get_db() -> Generator:
-    try:
-        db = SessionLocal()
-        yield db
-    finally:
-        db.close()
