+++ mcp-api-backend/utils/utils.py
@@ -1,0 +1,4 @@
+from sqlalchemy import inspect
+
+def object_as_dict(obj):
+    return {c.key: getattr(obj, c.key) for c in inspect(obj).mapper.column_attrs}
